requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.14.0
billiard==4.2.0
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-js-asset==2.2.0
django-nested-admin==4.1.0
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
easy-thumbnails==2.8.5
exceptiongroup==1.2.1
executing==2.0.1
flower==2.0.1
fonttools==4.53.1
frozenlist==1.5.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
ipython==8.20.0
jedi==0.19.1
jsmin==3.0.1
kombu==5.3.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
multidict==6.1.0
netaddr==1.3.0
oscrypto==1.3.0
parso==0.8.3
pdf2image==1.17.0
pexpect==4.9.0
Pillow==10.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-moneyed==3.0
pycparser==2.21
pydyf==0.11.0
Pygments==2.17.2
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyMuPDF==1.24.14
pypdf==5.1.0
pyphen==0.15.0
pyTelegramBotAPI==4.24.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.2
qrcode==8.0
redis==5.0.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.18.0
yookassa==3.3.0
zope.event==5.0
zope.interface==6.2
zopfli==0.2.3
